BaseActivity created for fancier code
